fix : 댓글 삭제 로직 수정 댓글 삭제시 confirm창 오픈, 삭제한다면 페이지 번호 1로 이동 및 페이지 번호 수정

+++ src/components/CommentList.js
@@ -44,6 +44,8 @@ function CommentList({ comments }) {
   const pageNumber = useSelector(state => state.page)
 
   const deleteHandler = (id) => {
+    const confirm = window.confirm("삭제 하시겠습니까?") 
+    if(!confirm) return
     dispatch(deleteComment(id))
     dispatch(selectPage(0))
   }
